chore: refactor anon Fx to tabKey Fx. Closes #14
